add passing of command line args to readme file

+++ sdk/device.py
@@ -9,8 +9,11 @@ class Device(object):
     refreshToken = None
     languageKey = "en"
     DB = "/Users/rdottin/Documents/Personal/pythonstarships/collectallresources/.device"
+    authentication_string = None
 
-    def __init__(self, name="DeviceTypeMac", key=None, language="en"):
+    def __init__(
+        self, name="DeviceTypeMac", key=None, language="en", authentication_string=None
+    ):
 
         if not key:
             key = "{}-{}-{}-{}-{}".format(
@@ -24,6 +27,9 @@ class Device(object):
         self.name = name
         self.key = key
         self.languageKey = language
+        self.authentication_string = authentication_string
+        if authentication_string:
+            self.DB = None
 
         if not self.load():
             self.save()
@@ -34,25 +40,28 @@ class Device(object):
         self.save()
 
     def reset(self):
-
-        os.unlink(self.DB)
+        if self.DB:
+            os.unlink(self.DB)
 
     def save(self):
-
-        with open(self.DB, "w+") as f:
-            f.write(
-                "{}|{}|{}|{}".format(
-                    self.name, self.key, self.refreshToken, self.languageKey
+        if self.DB:
+            with open(self.DB, "w+") as f:
+                f.write(
+                    "{}|{}|{}|{}".format(
+                        self.name, self.key, self.refreshToken, self.languageKey
+                    )
                 )
-            )
 
     def load(self):
+        if self.authentication_string:
+            data = self.authentication_string.split("|")
 
-        if not os.path.isfile(self.DB):
+        elif not os.path.isfile(self.DB):
             return False
 
-        with open(self.DB, "r") as f:
-            data = f.read().split("|")
+        else:
+            with open(self.DB, "r") as f:
+                data = f.read().split("|")
 
         self.name = data[0]
         self.key = data[1]
